Move all config types and predicates to a separate file
